Add bootstrap and organize folders
