--- components/ListDetail.tsx
@@ -1,16 +1,0 @@
-import * as React from 'react'
-
-import { User } from '../interfaces'
-
-type ListDetailProps = {
-  item: User
-}
-
-const ListDetail = ({ item: user }: ListDetailProps) => (
-  <div>
-    <h1>Detail for {user.name}</h1>
-    <p>ID: {user.id}</p>
-  </div>
-)
-
-export default ListDetail
