+++ package.json
@@ -11,12 +11,16 @@
     "axios": "^0.21.1",
     "next": "latest",
     "react": "^17.0.1",
-    "react-dom": "^17.0.1"
+    "react-dom": "^17.0.1",
   },
   "devDependencies": {
+    "@types/bootstrap": "^5.0.16",
     "@types/node": "^12.12.21",
     "@types/react": "^17.0.2",
+    "@types/react-bootstrap": "^0.32.25",
     "@types/react-dom": "^17.0.1",
+    "bootstrap": "^5.0.2",
+    "react-bootstrap": "^1.6.1",
     "typescript": "4.0"
   },
   "license": "MIT"
