+++ src/components/Layout/Layout.module.css
@@ -1,0 +1,5 @@
+.container {
+    max-width: 36rem;
+    padding: 0 1rem;
+    margin: 3rem auto 6rem;
+  }
