+++ src/pages/index.tsx
@@ -1,0 +1,10 @@
+import Layout from '../components/Layout/Layout'
+import Homepage from './homepage/homepage'
+
+const IndexPage = () => (
+  <Layout title="Vacinow">
+    <Homepage />
+  </Layout>
+)
+
+export default IndexPage
