+++ src/pages/upload/upload.tsx
@@ -1,6 +1,6 @@
 import { ChangeEvent, FormEvent, useState } from "react";
-import Layout from "../components/Layout";
-import api from "../services/api";
+import Layout from "../../components/Layout/Layout";
+import api from "../../services/api";
 
 const Upload = () => {
